products page should be products.js
