+++ app/server/schema.graphql
@@ -2066,6 +2066,29 @@ enum FormResultsOrderBy {
   PRIMARY_KEY_DESC
 }
 
+"""All input for the `insertDummyApplicationData` mutation."""
+input InsertDummyApplicationDataInput {
+  """
+  An arbitrary string value with no semantic meaning. Will be included in the
+  payload verbatim. May be used to track mutations by the client.
+  """
+  clientMutationId: String
+}
+
+"""The output of our `insertDummyApplicationData` mutation."""
+type InsertDummyApplicationDataPayload {
+  """
+  The exact same `clientMutationId` that was provided in the mutation input,
+  unchanged and unused. May be used by a client to track mutations.
+  """
+  clientMutationId: String
+
+  """
+  Our root query field type. Allows us to run any query from our mutation payload.
+  """
+  query: Query
+}
+
 """
 A JavaScript object encoded in the JSON format as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf).
 """
@@ -2318,6 +2341,12 @@ type Mutation {
     """
     input: DeleteUserByRowIdInput!
   ): DeleteUserPayload
+  insertDummyApplicationData(
+    """
+    The exclusive input argument for this mutation. An object type, make sure to see documentation for this object’s fields.
+    """
+    input: InsertDummyApplicationDataInput!
+  ): InsertDummyApplicationDataPayload
 }
 
 """An object with a globally unique `ID`."""
